# By 应雨洋
# Via 应雨洋
* 'temp-2.7.2' of github.com:LoyoChengdu/loyo-android:
  客户下拜访列表，线索下跟进
  客户详情下的拜访列表 完成

+++ app/src/main/java/com/loyo/oa/v2/activityui/signinnew/model/SigninNewListModel.java
@@ -25,7 +25,7 @@ public class SigninNewListModel implements Serializable {
     public String attachmentUUId; //uuid
     public String memo;           //内容
     public String atNameAndDepts; //内容
-    public float offsetDistance;  //偏差距离
+    public double offsetDistance;  //偏差距离
     public ArrayList<AudioModel>  audioInfo;       //语音录音
     public ArrayList<Attachment> attachments;      //文件
     public ArrayList<Attachment> imageAttachments; //图片
